Got both the HSH and PTM fitters to work when run from the CropExecuteLayout. Implemented the back button in CropExecuteLayout. Added the fitter output to the text area in CropExecuteLayout.
Created the HighlightDetectionLayout class, which is the layout for the
highlight detection process, and the HighlightDetectionListener, which
listens for events from the layout. Created the GUI for the layout,
and implemented it in main, so it can now bea reached from the
NewProjectLayout when the current project is a highlight type.
Implemented the handing of images from the NewProjectLayout to the
HighlightDetectionLayout, for both jpegs and non-jpegs.

Tomorrow, start work on implementing the highlight detection.