Added functionality to open projects using LP files with full image paths:  - added a radio button for this on the initial layout  - added a new dialog: LoadExistingLPDialog to handle opening this file  - added functions in dialog to open images in folder and pass to    CropExecuteLayout  - implemented back button for CropExecuteLayout to take program back    to initial screen when suing an LP file with full paths
Added function to filter out any images where an NaN/infinity was
generated for the highlight pos on that image, when using highlight
detection. A popup will show if this is the case.

Added check in redLPFile function to make sure there are no spaces in
image paths in LP files.